refactor: extract admin activity projection
